# perf: Moved global style creation outside of the React render cycle